+++ hydra-cluster/test/Test/EndToEndSpec.hs
@@ -82,6 +82,7 @@ import Hydra.Cluster.Scenarios (
   startWithWrongPeers,
   threeNodesNoErrorsOnOpen,
   threeNodesWithMirrorParty,
+  waitsForChainInSynchAndSecure,
  )
 import Hydra.Cluster.Util (chainConfigFor, keysFor, modifyConfig)
 import Hydra.Ledger.Cardano (mkRangedTx, mkSimpleTx)
@@ -100,6 +101,7 @@ import HydraNode (
   waitFor,
   waitForAllMatch,
   waitForNodesConnected,
+  waitForNodesSynced,
   waitMatch,
   withHydraCluster,
   withHydraNode,
@@ -405,6 +407,7 @@ spec = around (showLogsOnFailure "EndToEndSpec") $ do
               let hydraTracer = contramap FromHydraNode tracer
 
               withHydraCluster hydraTracer tmpDir nodeSocket' firstNodeId cardanoKeys hydraKeys hydraScriptsTxId contestationPeriod $ \nodes -> do
+                waitForNodesSynced hydraTracer backend (toList nodes)
                 waitForNodesConnected hydraTracer 20 nodes
                 let [n1, n2, n3] = toList nodes
 
@@ -413,6 +416,7 @@ spec = around (showLogsOnFailure "EndToEndSpec") $ do
                 seedFromFaucet_ backend bobCardanoVk 100_000_000 (contramap FromFaucet tracer)
                 seedFromFaucet_ backend carolCardanoVk 100_000_000 (contramap FromFaucet tracer)
 
+                -- Init head
                 send n1 $ input "Init" []
                 headId <-
                   waitForAllMatch 10 [n1, n2, n3] $ headIsInitializingWith (Set.fromList [alice, bob, carol])
@@ -470,6 +474,7 @@ spec = around (showLogsOnFailure "EndToEndSpec") $ do
             let hydraTracer = contramap FromHydraNode tracer
 
             withHydraCluster hydraTracer tmpDir nodeSocket' firstNodeId cardanoKeys hydraKeys hydraScriptsTxId contestationPeriod $ \nodes -> do
+              waitForNodesSynced hydraTracer backend (toList nodes)
               waitForNodesConnected hydraTracer 20 nodes
               let [n1, n2, n3] = toList nodes
 
@@ -601,6 +606,7 @@ spec = around (showLogsOnFailure "EndToEndSpec") $ do
             (tip, aliceHeadId) <- withHydraNode hydraTracer aliceChainConfig tmp nodeId aliceSk [] [1] $ \n1 -> do
               seedFromFaucet_ backend aliceCardanoVk 100_000_000 (contramap FromFaucet tracer)
               tip <- Backend.queryTip backend
+              waitForNodesSynced hydraTracer backend [n1]
               send n1 $ input "Init" []
               headId <- waitForAllMatch 10 [n1] $ headIsInitializingWith (Set.fromList [alice])
               return (tip, headId)
@@ -645,8 +651,8 @@ spec = around (showLogsOnFailure "EndToEndSpec") $ do
 
             withAliceNode $ \n1 -> do
               headId <- withBobNode $ \n2 -> do
+                waitForNodesSynced hydraTracer backend [n1, n2]
                 waitForNodesConnected hydraTracer 20 $ n1 :| [n2]
-
                 send n1 $ input "Init" []
                 headId <- waitForAllMatch 10 [n1, n2] $ headIsInitializingWith (Set.fromList [alice, bob])
 
@@ -713,6 +719,12 @@ spec = around (showLogsOnFailure "EndToEndSpec") $ do
             publishHydraScriptsAs backend Faucet
               >>= canResumeOnMemberAlreadyBootstrapped tracer tmpDir backend
 
+      it "prevent network interactions until chain backend is in sync and secure." $ \tracer -> do
+        withClusterTempDir $ \tmpDir -> do
+          withCardanoNodeDevnet (contramap FromCardanoNode tracer) tmpDir $ \_ backend ->
+            publishHydraScriptsAs backend Faucet
+              >>= waitsForChainInSynchAndSecure tracer tmpDir backend
+
     describe "two hydra heads scenario" $ do
       it "two heads on the same network do not conflict" $ \tracer ->
         failAfter 60 $
@@ -754,6 +766,7 @@ spec = around (showLogsOnFailure "EndToEndSpec") $ do
                   seedFromFaucet_ backend aliceCardanoVk 100_000_000 (contramap FromFaucet tracer)
                   seedFromFaucet_ backend bobCardanoVk 100_000_000 (contramap FromFaucet tracer)
 
+                  waitForNodesSynced hydraTracer backend [n1, n2]
                   send n1 $ input "Init" []
                   headIdAliceOnly <- waitMatch 10 n1 $ headIsInitializingWith (Set.fromList [alice])
 
@@ -788,6 +801,7 @@ spec = around (showLogsOnFailure "EndToEndSpec") $ do
                   withHydraNode hydraTracer carolChainConfig tmpDir 3 carolSk [aliceVk, bobVk] allNodeIds $ \n3 -> do
                     -- Funds to be used as fuel by Hydra protocol transactions
                     seedFromFaucet_ backend aliceCardanoVk 100_000_000 (contramap FromFaucet tracer)
+                    waitForNodesSynced hydraTracer backend [n1, n2, n3]
                     waitForNodesConnected hydraTracer 20 $ n1 :| [n2, n3]
                     send n1 $ input "Init" []
                     void $ waitForAllMatch 3 [n1] $ headIsInitializingWith (Set.fromList [alice, bob, carol])
@@ -856,6 +870,7 @@ spec = around (showLogsOnFailure "EndToEndSpec") $ do
             let contestationPeriod = 2
             aliceChainConfig <- chainConfigFor Alice dir backend hydraScriptsTxId [] contestationPeriod
             withHydraNode hydraTracer aliceChainConfig dir 1 aliceSk [] [1] $ \n1 -> do
+              waitForNodesSynced hydraTracer backend [n1]
               send n1 $ input "Init" []
 
             let logFilePath = dir </> "logs" </> "hydra-node-1.log"
@@ -874,6 +889,7 @@ timedTx tmpDir tracer backend hydraScriptsTxId = do
 
     -- Funds to be used as fuel by Hydra protocol transactions
     seedFromFaucet_ backend aliceCardanoVk lovelaceBalanceValue (contramap FromFaucet tracer)
+    waitForNodesSynced hydraTracer backend [n1]
     send n1 $ input "Init" []
     headId <-
       waitForAllMatch 10 [n1] $
@@ -941,6 +957,7 @@ initAndClose tmpDir tracer clusterIx hydraScriptsTxId backend = do
         Direct DirectOptions{nodeSocket} -> nodeSocket
         _ -> error "Unexpected Blockfrost backend"
   withHydraCluster hydraTracer tmpDir nodeSocket' firstNodeId cardanoKeys hydraKeys hydraScriptsTxId contestationPeriod $ \nodes -> do
+    waitForNodesSynced hydraTracer backend (toList nodes)
     let [n1, n2, n3] = toList nodes
     waitForNodesConnected hydraTracer 20 $ n1 :| [n2, n3]
 
@@ -1057,8 +1074,8 @@ reachFanoutLimit ledgerSize tmpDir tracer hydraScriptsTxId backend = do
         _ -> error "Unexpected Blockfrost backend"
 
   withHydraCluster hydraTracer tmpDir nodeSocket' 1 [aliceKeys] [aliceSk] hydraScriptsTxId contestationPeriod $ \nodes -> do
+    waitForNodesSynced hydraTracer backend (toList nodes)
     let [node] = toList nodes
-
     waitForNodesConnected hydraTracer 20 $ node :| []
 
     -- Funds to be used as fuel by Hydra protocol transactions