+++ hydra-cluster/test/Test/Hydra/Cluster/HydraClientSpec.hs
@@ -53,6 +53,7 @@ import HydraNode (
   waitFor,
   waitForAllMatch,
   waitForNodesConnected,
+  waitForNodesSynced,
   waitMatch,
   waitNoMatch,
   withConnectionToNodeHost,
@@ -265,6 +266,7 @@ scenarioSetup tracer tmpDir action = do
     let hydraTracer = contramap FromHydraNode tracer
 
     withHydraCluster hydraTracer tmpDir nodeSocket' firstNodeId cardanoKeys hydraKeys hydraScriptsTxId contestationPeriod $ \nodes -> do
+      waitForNodesSynced hydraTracer backend (toList nodes)
       let [n1, n2, n3] = toList nodes
       waitForNodesConnected hydraTracer 20 $ n1 :| [n2, n3]