+++ hydra-tui/src/Hydra/TUI/Handlers.hs
@@ -131,6 +131,10 @@ handleHydraEventsConnection now = \case
   Update (ApiTimedServerOutput TimedServerOutput{output = API.NetworkDisconnected}) -> do
     networkStateL .= Just NetworkDisconnected
     peersL %= map (\(h, _) -> (h, PeerIsUnknown))
+  Update (ApiTimedServerOutput TimedServerOutput{time, output = API.NodeUnsynced}) -> do
+    chainSyncedStatusL .= OutOfSync
+  Update (ApiTimedServerOutput TimedServerOutput{time, output = API.NodeSynced}) -> do
+    chainSyncedStatusL .= InSync
   e -> zoom headStateL $ handleHydraEventsHeadState now e
  where
   updatePeerStatus :: Host -> PeerStatus -> [(Host, PeerStatus)] -> [(Host, PeerStatus)]
@@ -228,6 +232,13 @@ handleHydraEventsInfo = \case
   Update (ApiClientMessage API.CommandFailed{clientInput}) -> do
     time <- liftIO getCurrentTime
     warn time $ "Invalid command: " <> show clientInput
+  Update (ApiClientMessage API.RejectedInput{clientInput, reason}) -> do
+    time <- liftIO getCurrentTime
+    warn time $ "Rejected command: " <> show clientInput <> " Reason: " <> show reason
+  Update (ApiTimedServerOutput TimedServerOutput{time, output = API.NodeUnsynced}) -> do
+    warn time "Node state is out of sync with chain backend."
+  Update (ApiTimedServerOutput TimedServerOutput{time, output = API.NodeSynced}) -> do
+    warn time "Node state is back in sync with chain backend."
   Update (ApiTimedServerOutput TimedServerOutput{time, output = API.HeadIsClosed{snapshotNumber}}) ->
     info time $ "Head closed with snapshot number " <> show snapshotNumber
   Update (ApiTimedServerOutput TimedServerOutput{time, output = API.HeadIsContested{snapshotNumber, contestationDeadline}}) ->