+++ hydra-tui/test/Hydra/TUISpec.hs
@@ -47,7 +47,7 @@ import Hydra.Options (DirectOptions (..), RunOptions, persistenceRotateAfter)
 import Hydra.TUI (runWithVty)
 import Hydra.TUI.Drawing (renderTime)
 import Hydra.TUI.Options (Options (..))
-import Hydra.Tx.ContestationPeriod (ContestationPeriod, toNominalDiffTime)
+import Hydra.Tx.ContestationPeriod (ContestationPeriod, toNominalDiffTime, unsyncedPolicy)
 import HydraNode (
   HydraClient (HydraClient, hydraNodeId),
   HydraNodeLog,
@@ -205,6 +205,22 @@ spec = do
       let time' = 1 * hours + 1 * minutes + 15 * seconds
       renderTime (-time' :: NominalDiffTime) `shouldBe` "-0d 1h 1m 15s"
 
+    around setupRotatedStateTUI $ do
+      it "should show the chain synced status" $
+        \TUIRotatedTest
+          { tuiTest = TUITest{shouldRender}
+          , nodeHandle = HydraNodeHandle{stopNode, startNode}
+          , blockTime
+          } -> do
+            threadDelay 1
+            shouldRender "Synced"
+            stopNode
+            -- Wait for some blocks to roll forward
+            threadDelay $ realToFrac (unsyncedPolicy tuiContestationPeriod + 10 * blockTime)
+            startNode
+            threadDelay 1
+            shouldRender "CatchingUp"
+
   context "text rendering errors" $ do
     around setupNotEnoughFundsNodeAndTUI $ do
       it "should show not enough fuel message and suggestion" $
@@ -220,7 +236,7 @@ setupRotatedStateTUI :: (TUIRotatedTest -> IO ()) -> IO ()
 setupRotatedStateTUI action = do
   showLogsOnFailure "TUISpec" $ \tracer ->
     withTempDir "tui-end-to-end" $ \tmpDir -> do
-      withCardanoNodeDevnet (contramap FromCardano tracer) tmpDir $ \_ backend -> do
+      withCardanoNodeDevnet (contramap FromCardano tracer) tmpDir $ \blockTime backend -> do
         hydraScriptsTxId <- publishHydraScriptsAs backend Faucet
         chainConfig <- chainConfigFor Alice tmpDir backend hydraScriptsTxId [] tuiContestationPeriod
         let nodeId = 1
@@ -232,11 +248,12 @@ setupRotatedStateTUI action = do
         seedFromFaucet_ backend aliceCardanoVk 100_000_000 (contramap FromFaucet tracer)
         options <- prepareHydraNode chainConfig tmpDir nodeId aliceSk [] [nodeId] id
         let options' = options{persistenceRotateAfter = Just (Positive 1)}
-        withTUIRotatedTest (contramap FromHydra tracer) tmpDir nodeId backend externalKeyFilePath options' action
+        withTUIRotatedTest (contramap FromHydra tracer) tmpDir nodeId blockTime backend externalKeyFilePath options' action
 
 data TUIRotatedTest = TUIRotatedTest
   { tuiTest :: TUITest
   , nodeHandle :: HydraNodeHandle
+  , blockTime :: NominalDiffTime
   }
 
 data HydraNodeHandle = HydraNodeHandle
@@ -287,13 +304,14 @@ withTUIRotatedTest ::
   Tracer IO HydraNodeLog ->
   FilePath ->
   Int ->
+  NominalDiffTime ->
   DirectBackend ->
   FilePath ->
   RunOptions ->
   (TUIRotatedTest -> Expectation) ->
   Expectation
-withTUIRotatedTest tracer tmpDir nodeId backend externalKeyFilePath options' action = do
-  withHydraNodeHandle tracer tmpDir nodeId options' $ \nodeHandle -> do
+withTUIRotatedTest tracer tmpDir nodeId blockTime backend externalKeyFilePath options action =
+  withHydraNodeHandle tracer tmpDir nodeId options $ \nodeHandle -> do
     startNode nodeHandle
     HydraClient{hydraNodeId} <- getClient nodeHandle
     withTUITest (150, 10) $ \brickTest@TUITest{buildVty} -> do
@@ -320,6 +338,7 @@ withTUIRotatedTest tracer tmpDir nodeId backend externalKeyFilePath options' act
             TUIRotatedTest
               { tuiTest = brickTest
               , nodeHandle
+              , blockTime
               }
         )
  where