+++ hydra-tx/src/Hydra/Tx/ContestationPeriod.hs
@@ -86,3 +86,9 @@ fromChain cp =
   UnsafeContestationPeriod
     . truncate
     $ toInteger (OnChain.milliseconds cp) % 1000
+
+-- | Period of time after which we consider the node becoming unsynced with the chain.
+-- Beyond this period the node will refuse to process new transactions and signing snapshots.
+-- FIXME: this is implicit, consider making it configurable.
+unsyncedPolicy :: ContestationPeriod -> NominalDiffTime
+unsyncedPolicy cp = toNominalDiffTime cp * 0.5