+++ CHANGELOG.md
@@ -13,6 +13,10 @@ changes.
 
 - `hydra-node` has a new endpoint `GET /head-initialization` which serves the timestamp of the last Head initialization.
 
+- **BREAKING** Hydra node stops interacting on network and client API once the node's view of the chain is out of sync more than 50% of `--contestation-period`.
+  - Added `NodeUnsynced` and `NodeSynced` state events and server outputs.
+  - Added `RejectedInput` client message.
+  - The `Checkpoint` event, and consequently the `EventLogRotated` server output, now carry the different `NodeState` variants: `NodeInSync` or `NodeCatchingUp`.
 
 ## [1.1.0] - 2025-10-28